tobati: Dowgrade GPU
FirePro W5000 has a proper DVI output, and is cheaper than an
active DP -> dual-link DVI adapter.

I'm not sure why the amdgpu driver doesn't work with it, but
switching was a bit awkward because amdgpu blacklists the
radeon kernel module, and I had to manually rmmod amdgpu when
switching.