wip: install-scripts: More automation and separation of parts
part-zfs-boot.sh on a 1TB drive used GPT where Grub was unable to embed itself.
and part-boot-root-rest.sh let the install succeed, but it was then unbootable. This
was on an SSD that previously had Windows, and I suspect some part of the Windows
boot loader wasn't being properly wiped.

Also switch to encrypted ZFS by default, match volume name case to
mountpoints and add tmp since I'm not using tmp on tmpfs as much
as when I first wrote format-zfs.sh